Задание 2/Код/requirements.txt
allure-pytest
pytest
selenium
webdriver-manager